Tried shared-bus-rtic
This exercise made me realize that with the current shared bus architecture, we
will not be able to concurrently service USB interrupts and write to the EPD
display.  This is because the USB interrupt needs access to flash, which is in
the same bus as the EPD.  Instead of juggling crazy locking hacks, change the
hardware design to use dedicated SPI buses for each.

Checking this in which might to put it on stand by until new hardware arrives.

This does not compile.