Work around mismatch between flash and scsi block size
Implement a simple write function that erases sectors when writing to
the start of sector boundary.  This will not work for write operations
that are 1. unaligned with 4k sector boundary, and 2. that are not a
multiple of 4k bytes.  But that is all we need.

The implementation of read-modify-write is commented out as it was
resulting in large chunks of unwritten data.